implement stock import
