+++ todo.txt
@@ -28,6 +28,7 @@ MINOR IMPROVEMENTS
 - review endpoint, dates on each note, price of share at the time of the note, separate each note
 - notes, fill with share information when note is made
 - organize code: separate events from handlers from core domain models?
+- import validation (size, contents, headers)
 
 DESIGN
 - consider cards for rendering owned options, owned stock... might get too much?
